data to be balanced across classes/randomize the data
Merge and prune the training data as needed. Depending on your computer
setup, you might not be able to fit it all in memory, and you can tune
train_size as needed. The labels will be stored into a separate array of
integers 0 through 9.
Also create a validation dataset for hyperparameter tuning.

It's important to have the labels well shuffled for the training and
test distributions to match.